(fix) Enhance result form to handle non-set order concepts
