+++ src/work-list/work-list.component.tsx
@@ -47,7 +47,6 @@ import {
 import { launchOverlay } from "../components/overlay/hook";
 import ResultForm from "../results/result-form.component";
 import { getStatusColor } from "../utils/functions";
-import LabResultForm from "../results/single-result-form.component";
 
 interface WorklistProps {
   fulfillerStatus: string;
@@ -127,8 +126,7 @@ const WorkList: React.FC<WorklistProps> = ({ fulfillerStatus }) => {
         onClick={() => {
           launchOverlay(
             t("resultForm", "Result  Tests"),
-            // <ResultForm patientUuid={patientUuid} order={order} />
-            <LabResultForm patientUuid={patientUuid} order={order} />
+            <ResultForm patientUuid={patientUuid} order={order} />
           );
         }}
         renderIcon={(props) => <Microscope size={16} {...props} />}
